Manage product tag

+++ back_end/src/main/java/com/example/back_end/entity/LocalizedProperty.java
@@ -45,6 +45,6 @@ public class LocalizedProperty {
     private Language language;
 
     @Column(name = "entity_id")
-    private Integer entityId;
+    private Long entityId;
 
 }
